Fix logins not being associated with ids

+++ src/rest.rs
@@ -4,7 +4,7 @@ use crate::db::validate_login;
 use crate::db::{self, APPLICATION_ACCPETED, APPLICATION_DENIED, APPLICATION_PENDING, ID};
 use crate::db::{ApplicantIDNameField, DbConn};
 use crate::models::*;
-use crate::request_guards::Administrator;
+use crate::request_guards::{AdminOrApplicant, AdminOrProfessor, Administrator, LoggedIn};
 use crate::SessionTokenState;
 use chrono::{Duration, Local};
 use rand_chacha::rand_core::RngCore;
@@ -28,6 +28,7 @@ struct IdPayload {
 async fn create_research_field(
     conn: DbConn,
     research_field: Json<NewResearchField>,
+    _admin: Administrator,
 ) -> Result<Json<IdPayload>, Status> {
     match db::create_research_field(&conn, research_field.into_inner().name).await {
         Ok(id) => Ok(Json(IdPayload { id })),
@@ -45,7 +46,10 @@ async fn create_research_field(
 #[get("/research-field?<id>")]
 async fn get_research_field(conn: DbConn, id: i32) -> Result<Json<ResearchField>, Status> {
     match db::get_research_field(&conn, id).await {
-        Ok(research_field) => Ok(Json(research_field)),
+        Ok(research_field) => match research_field {
+            Some(research_field) => Ok(Json(research_field)),
+            None => Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error occured while trying to get research field: {}", e);
             Err(Status::InternalServerError)
@@ -55,7 +59,10 @@ async fn get_research_field(conn: DbConn, id: i32) -> Result<Json<ResearchField>
 
 /// Endpoint for getting all research fields.
 #[get("/research-fields")]
-async fn get_research_fields(conn: DbConn) -> Result<Json<Vec<ResearchField>>, Status> {
+async fn get_research_fields(
+    conn: DbConn,
+    _logged_in: LoggedIn,
+) -> Result<Json<Vec<ResearchField>>, Status> {
     match db::get_research_fields(&conn).await {
         Ok(research_fields) => Ok(Json(research_fields)),
         Err(e) => {
@@ -67,7 +74,7 @@ async fn get_research_fields(conn: DbConn) -> Result<Json<Vec<ResearchField>>, S
 
 /// Endpoint for deleting a research field.
 #[delete("/research-field?<id>")]
-async fn delete_research_field(conn: DbConn, id: i32) -> Status {
+async fn delete_research_field(conn: DbConn, id: i32, _admin: Administrator) -> Status {
     match db::delete_research_field(&conn, id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -85,6 +92,7 @@ async fn delete_research_field(conn: DbConn, id: i32) -> Status {
 async fn create_professor(
     conn: DbConn,
     professor: Json<NewProfessor>,
+    _admin: Administrator,
 ) -> Result<Json<IdPayload>, Status> {
     match db::create_professor(&conn, professor.into_inner().name).await {
         Ok(id) => Ok(Json(IdPayload { id })),
@@ -97,9 +105,16 @@ async fn create_professor(
 
 /// Endpoint for getting a professor.
 #[get("/professor?<id>")]
-async fn get_professor(conn: DbConn, id: i32) -> Result<Json<Professor>, Status> {
+async fn get_professor(
+    conn: DbConn,
+    id: i32,
+    _logged_in: LoggedIn,
+) -> Result<Json<Professor>, Status> {
     match db::get_professor(&conn, id).await {
-        Ok(professor) => Ok(Json(professor)),
+        Ok(professor) => match professor {
+            Some(professor) => Ok(Json(professor)),
+            None => Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error occured while trying to get professor: {}", e);
             Err(Status::InternalServerError)
@@ -109,7 +124,10 @@ async fn get_professor(conn: DbConn, id: i32) -> Result<Json<Professor>, Status>
 
 /// Endpoint for getting all professors.
 #[get("/professors")]
-async fn get_professors(conn: DbConn) -> Result<Json<Vec<Professor>>, Status> {
+async fn get_professors(
+    conn: DbConn,
+    _logged_in: LoggedIn,
+) -> Result<Json<Vec<Professor>>, Status> {
     match db::get_professors(&conn).await {
         Ok(v) => Ok(Json(v)),
         Err(e) => {
@@ -121,7 +139,7 @@ async fn get_professors(conn: DbConn) -> Result<Json<Vec<Professor>>, Status> {
 
 /// Endpoint for deleting a professor.
 #[delete("/professor?<id>")]
-async fn delete_professor(conn: DbConn, id: i32) -> Status {
+async fn delete_professor(conn: DbConn, id: i32, _admin: Administrator) -> Status {
     match db::delete_professor(&conn, id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -133,7 +151,16 @@ async fn delete_professor(conn: DbConn, id: i32) -> Status {
 
 /// Endpoint for adding a research field to a professor.
 #[post("/professor/research-field?<prof_id>&<field_id>")]
-async fn add_researched_field_to_professor(conn: DbConn, prof_id: i32, field_id: i32) -> Status {
+async fn add_researched_field_to_professor(
+    conn: DbConn,
+    prof_id: i32,
+    field_id: i32,
+    admin_or_professor: AdminOrProfessor,
+) -> Status {
+    if !admin_or_professor.can_access_prof(prof_id) {
+        return Status::Forbidden;
+    }
+
     match db::add_researched_field_to_professor(&conn, prof_id, field_id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -151,6 +178,7 @@ async fn add_researched_field_to_professor(conn: DbConn, prof_id: i32, field_id:
 async fn get_fields_professor_researches(
     conn: DbConn,
     prof_id: i32,
+    _logged_in: LoggedIn,
 ) -> Result<Json<Vec<ResearchField>>, Status> {
     match db::get_fields_professor_researches(&conn, prof_id).await {
         Ok(research_fields) => Ok(Json(research_fields)),
@@ -169,6 +197,7 @@ async fn get_applicants_for_professor_with_status(
     conn: DbConn,
     id: i32,
     status: String,
+    _logged_in: LoggedIn,
 ) -> Result<Json<Vec<ApplicantIDNameField>>, Status> {
     match status.as_str() {
         APPLICATION_ACCPETED => {}
@@ -201,7 +230,12 @@ async fn remove_researched_field_from_professor(
     conn: DbConn,
     prof_id: i32,
     field_id: i32,
+    admin_or_professor: AdminOrProfessor,
 ) -> Status {
+    if !admin_or_professor.can_access_prof(prof_id) {
+        return Status::Forbidden;
+    }
+
     match db::remove_researched_field_from_professor(&conn, prof_id, field_id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -219,7 +253,12 @@ pub async fn accept_application(
     conn: DbConn,
     applicant_id: i32,
     professor_id: i32,
+    admin_or_professor: AdminOrProfessor,
 ) -> Result<(), Status> {
+    if !admin_or_professor.can_access_prof(professor_id) {
+        return Err(Status::Forbidden);
+    }
+
     if let Err(e) = db::accept_applicant_application(&conn, applicant_id, professor_id).await {
         eprintln!("Error while accepting an applicant's application: {}", e);
         Err(Status::InternalServerError)
@@ -233,7 +272,12 @@ pub async fn deny_application(
     conn: DbConn,
     applicant_id: i32,
     professor_id: i32,
+    admin_or_professor: AdminOrProfessor,
 ) -> Result<(), Status> {
+    if !admin_or_professor.can_access_prof(professor_id) {
+        return Err(Status::Forbidden);
+    }
+
     if let Err(e) = db::deny_applicant_application(&conn, applicant_id, professor_id).await {
         eprintln!("Error while denying an applicant's application: {}", e);
         Err(Status::InternalServerError)
@@ -247,6 +291,7 @@ pub async fn deny_application(
 async fn create_applicant(
     conn: DbConn,
     applicant: Json<NewApplicant>,
+    _admin: Administrator,
 ) -> Result<Json<IdPayload>, Status> {
     match db::create_applicant(&conn, applicant.into_inner()).await {
         Ok(id) => Ok(Json(IdPayload { id })),
@@ -263,7 +308,12 @@ async fn edit_applicant(
     conn: DbConn,
     app_id: i32,
     applicant: Json<ApplicantEdit>,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Result<(), Status> {
+    if !admin_or_applicant.can_access_applicant(app_id) {
+        return Err(Status::Forbidden);
+    }
+
     match db::edit_applicant(&conn, app_id, applicant.into_inner()).await {
         Ok(_) => Ok(()),
         Err(e) => {
@@ -278,7 +328,12 @@ async fn edit_professor(
     conn: DbConn,
     prof_id: i32,
     professor: Json<ProfessorEdit>,
+    admin_or_professor: AdminOrProfessor,
 ) -> Result<(), Status> {
+    if !admin_or_professor.can_access_prof(prof_id) {
+        return Err(Status::Forbidden);
+    }
+
     match db::edit_professor(&conn, prof_id, professor.into_inner()).await {
         Ok(_) => Ok(()),
         Err(e) => {
@@ -290,9 +345,16 @@ async fn edit_professor(
 
 /// Endpoint for getting an applicant.
 #[get("/applicant?<id>")]
-async fn get_applicant(conn: DbConn, id: i32) -> Result<Json<Applicant>, Status> {
+async fn get_applicant(
+    conn: DbConn,
+    id: i32,
+    _logged_in: LoggedIn,
+) -> Result<Json<Applicant>, Status> {
     match db::get_applicant(&conn, id).await {
-        Ok(applicant) => Ok(Json(applicant)),
+        Ok(applicant) => match applicant {
+            Some(applicant) => Ok(Json(applicant)),
+            None => Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error occured while trying to get applicant: {}", e);
             Err(Status::InternalServerError)
@@ -302,7 +364,10 @@ async fn get_applicant(conn: DbConn, id: i32) -> Result<Json<Applicant>, Status>
 
 /// Endpoint for getting all applicants.
 #[get("/applicants")]
-async fn get_applicants(conn: DbConn) -> Result<Json<Vec<Applicant>>, Status> {
+async fn get_applicants(
+    conn: DbConn,
+    _logged_in: LoggedIn,
+) -> Result<Json<Vec<Applicant>>, Status> {
     match db::get_applicants(&conn).await {
         Ok(applicants) => Ok(Json(applicants)),
         Err(e) => {
@@ -314,7 +379,7 @@ async fn get_applicants(conn: DbConn) -> Result<Json<Vec<Applicant>>, Status> {
 
 /// Endpoint for deleting an applicant.
 #[delete("/applicant?<id>")]
-async fn delete_applicant(conn: DbConn, id: i32) -> Status {
+async fn delete_applicant(conn: DbConn, id: i32, _admin: Administrator) -> Status {
     match db::delete_applicant(&conn, id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -326,7 +391,16 @@ async fn delete_applicant(conn: DbConn, id: i32) -> Status {
 
 /// Endpoint for adding an application to an applicant.
 #[post("/applicant/applications?<applicant_id>&<prof_id>")]
-async fn add_application_to_applicant(conn: DbConn, applicant_id: i32, prof_id: i32) -> Status {
+async fn add_application_to_applicant(
+    conn: DbConn,
+    applicant_id: i32,
+    prof_id: i32,
+    admin_or_applicant: AdminOrApplicant,
+) -> Status {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Status::Forbidden;
+    }
+
     match db::add_application_to_applicant(&conn, applicant_id, prof_id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -349,7 +423,12 @@ async fn upload_applicant_cv(
     conn: DbConn,
     applicant_id: i32,
     file: Data<'_>,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Result<(), Status> {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Err(Status::Forbidden);
+    }
+
     let file = match file.open(get_file_upload_max()).into_bytes().await {
         Ok(v) => v,
         Err(e) => {
@@ -385,7 +464,12 @@ async fn upload_applicant_diploma(
     conn: DbConn,
     applicant_id: i32,
     file: Data<'_>,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Result<(), Status> {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Err(Status::Forbidden);
+    }
+
     let file = match file.open(get_file_upload_max()).into_bytes().await {
         Ok(v) => v,
         Err(e) => {
@@ -421,7 +505,12 @@ async fn upload_applicant_grade_audit(
     conn: DbConn,
     applicant_id: i32,
     file: Data<'_>,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Result<(), Status> {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Err(Status::Forbidden);
+    }
+
     let file = match file.open(get_file_upload_max()).into_bytes().await {
         Ok(v) => v,
         Err(e) => {
@@ -452,9 +541,16 @@ async fn upload_applicant_grade_audit(
 }
 
 #[get("/applicant/files/cv?<applicant_id>")]
-async fn get_applicant_cv(conn: DbConn, applicant_id: i32) -> Result<Vec<u8>, Status> {
+async fn get_applicant_cv(
+    conn: DbConn,
+    applicant_id: i32,
+    _logged_in: LoggedIn,
+) -> Result<Vec<u8>, Status> {
     let applicant = match db::get_applicant(&conn, applicant_id).await {
-        Ok(v) => v,
+        Ok(v) => match v {
+            Some(v) => v,
+            None => return Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error while fetching applicant: {}", e);
             return Err(Status::InternalServerError);
@@ -473,9 +569,16 @@ async fn get_applicant_cv(conn: DbConn, applicant_id: i32) -> Result<Vec<u8>, St
 }
 
 #[get("/applicant/files/diploma?<applicant_id>")]
-async fn get_applicant_diploma(conn: DbConn, applicant_id: i32) -> Result<Vec<u8>, Status> {
+async fn get_applicant_diploma(
+    conn: DbConn,
+    applicant_id: i32,
+    _logged_in: LoggedIn,
+) -> Result<Vec<u8>, Status> {
     let applicant = match db::get_applicant(&conn, applicant_id).await {
-        Ok(v) => v,
+        Ok(v) => match v {
+            Some(v) => v,
+            None => return Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error while fetching applicant: {}", e);
             return Err(Status::InternalServerError);
@@ -494,9 +597,16 @@ async fn get_applicant_diploma(conn: DbConn, applicant_id: i32) -> Result<Vec<u8
 }
 
 #[get("/applicant/files/grade-audit?<applicant_id>")]
-async fn get_applicant_grade_audit(conn: DbConn, applicant_id: i32) -> Result<Vec<u8>, Status> {
+async fn get_applicant_grade_audit(
+    conn: DbConn,
+    applicant_id: i32,
+    _logged_in: LoggedIn,
+) -> Result<Vec<u8>, Status> {
     let applicant = match db::get_applicant(&conn, applicant_id).await {
-        Ok(v) => v,
+        Ok(v) => match v {
+            Some(v) => v,
+            None => return Err(Status::NotFound),
+        },
         Err(e) => {
             eprintln!("DB error while fetching applicant: {}", e);
             return Err(Status::InternalServerError);
@@ -519,7 +629,12 @@ async fn get_applicant_grade_audit(conn: DbConn, applicant_id: i32) -> Result<Ve
 async fn get_profs_applicant_applied_to(
     conn: DbConn,
     applicant_id: i32,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Result<Json<Vec<Professor>>, Status> {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Err(Status::Forbidden);
+    }
+
     match db::get_profs_applicant_applied_to(&conn, applicant_id).await {
         Ok(professors) => Ok(Json(professors)),
         Err(e) => {
@@ -538,7 +653,12 @@ async fn remove_application_from_applicant(
     conn: DbConn,
     applicant_id: i32,
     prof_id: i32,
+    admin_or_applicant: AdminOrApplicant,
 ) -> Status {
+    if !admin_or_applicant.can_access_applicant(applicant_id) {
+        return Status::Forbidden;
+    }
+
     match db::remove_application_from_applicant(&conn, applicant_id, prof_id).await {
         Ok(_) => Status::Ok,
         Err(e) => {
@@ -574,7 +694,7 @@ pub async fn login(
     login_data: Json<Login>,
     session_tokens: &State<SessionTokenState>,
     cookies: &CookieJar<'_>,
-) -> Redirect {
+) -> Status {
     match validate_login(
         &conn,
         login_data.username.clone(),
@@ -592,10 +712,10 @@ pub async fn login(
             let expiry = Local::now() + Duration::days(30);
 
             session_tokens.insert(token, (session_type, expiry));
-            Redirect::to(uri!("/"))
+            Status::Ok
         }
         Err(e) => match e {
-            _ => Redirect::to(uri!("/login?error=true")),
+            _ => Status::Forbidden,
         },
     }
 }
@@ -639,13 +759,14 @@ pub async fn create_admin_login(
     }
 }
 
-#[post("/login/applicant", data = "<login_data>")]
+#[post("/login/applicant?<applicant_id>", data = "<login_data>")]
 pub async fn create_applicant_login(
     conn: DbConn,
     login_data: Json<Login>,
+    applicant_id: i32,
     _administrator: Administrator,
 ) -> Status {
-    match db::create_applicant_account(&conn, login_data.into_inner()).await {
+    match db::create_applicant_account(&conn, applicant_id, login_data.into_inner()).await {
         Ok(_) => Status::Ok,
         Err(e) => {
             eprintln!(
@@ -657,13 +778,14 @@ pub async fn create_applicant_login(
     }
 }
 
-#[post("/login/professor", data = "<login_data>")]
+#[post("/login/professor?<professor_id>", data = "<login_data>")]
 pub async fn create_professor_login(
     conn: DbConn,
     login_data: Json<Login>,
+    professor_id: i32,
     _administrator: Administrator,
 ) -> Status {
-    match db::create_professor_account(&conn, login_data.into_inner()).await {
+    match db::create_professor_account(&conn, professor_id, login_data.into_inner()).await {
         Ok(_) => Status::Ok,
         Err(e) => {
             eprintln!(
